refactor : (로또 게임) 총 상금 계산로직 리팩토링
